requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Flask-Cors==5.0.0
